llm/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.0.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.1
beartype==0.15.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
bokeh==3.3.4
boto3==1.34.29
botocore==1.34.29
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.0
contourpy==1.2.0
cryptography==42.0.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
diskcache==5.6.3
distro==1.9.0
entrypoints==0.4
executing==2.0.1
fastapi==0.109.0
fastjsonschema==2.19.0
filelock==3.13.1
fonttools==4.45.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.12.2
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.41
# google-api-core==2.15.0
# google-api-python-client==2.115.0
# google-auth==2.27.0
# google-auth-httplib2==0.2.0
# googleapis-common-protos==1.62.0
h11==0.14.0
h3==3.7.6
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
huggingface-hub==0.20.3
idna==3.4
importlib-metadata==7.0.1
ipykernel==6.26.0
ipython==8.17.2
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter-bokeh==3.0.7
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter_client==7.4.9
jupyter_core==5.5.0
jupyter_server==2.10.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.2
kiwisolver==1.4.5
linkify-it-py==2.0.2
litellm==1.20.0
llama_cpp_python==0.2.33
llvmlite==0.41.1
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
msal==1.26.0
multidict==6.0.4
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
notebook==7.0.6
notebook_shim==0.2.3
numba==0.58.1
numpy==1.26.2
Office365-REST-Python-Client==2.5.5
openai==1.10.0
overrides==7.4.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.0
panel==1.3.8
param==2.0.2
parso==0.8.3
pathway
pexpect==4.8.0
Pillow==10.1.0
platformdirs==4.0.0
pptree==3.1
prettyprint==0.1.5
prometheus-client==0.18.0
prompt-toolkit==3.0.40
protobuf==4.25.2
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-sat==0.1.8.dev13
pytz==2023.3.post1
pyviz_comms==3.0.1
PyYAML==6.0.1
pyzmq==24.0.1
referencing==0.31.0
regex==2023.12.25
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.13.1
rsa==4.9
s3transfer==0.10.0
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.12.0
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
shapely==2.0.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
sqlglot==10.6.1
sse-starlette==2.0.0
stack-data==0.6.3
starlette==0.35.1
starlette-context==0.3.6
streamlit==1.30.0
sympy==1.12
tenacity==8.2.3
terminado==0.18.0
threadpoolctl==3.2.0
tiktoken==0.5.2
tinycss2==1.2.1
tokenizers==0.15.1
toml==0.10.2
toolz==0.12.1
torch==2.1.2
torchvision==0.16.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
transformers==4.37.1
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
uc-micro-py==1.0.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.0.7
uvicorn==0.27.0
validators==0.22.0
wcwidth==0.2.10
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
widgetsnbextension==4.0.9
xyzservices==2023.10.1
yarl==1.9.4
zipp==3.17.0